requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
pydantic==2.5.3
pandas==2.1.4
numpy==1.26.3
yfinance==0.2.36
python-multipart==0.0.6
requests==2.31.0
fredapi==0.5.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2
